re organize some functions
